feat: add mock data for meetings by month and update dashboard components layout
