+++ src/pages/DashboardPage/SalesmanDashboard/SalesmanDashboard.tsx
@@ -1,6 +1,6 @@
 import { EpageDescriptions } from '@data/enums/EpageDescriptions';
 import { EPageTitles } from '@data/enums/EPageTitles';
-import { Stack } from '@mui/material';
+import { Box, Stack } from '@mui/material';
 import { AvgDurationLineChart } from '@pages/DashboardPage/components/AvgDurationLineChart/AvgDurationLineChart';
 import { MeetingsByMonthChart } from '@pages/DashboardPage/components/MeetingsByMonthChart/MeetingsByMonthChart';
 import { PageContainter } from '@UI/PageContainer/PageContainer';
@@ -15,8 +15,14 @@ export const SalesmanDashboard = (): JSX.Element => {
           title={EPageTitles.SALESMAN_DASHBOARD}
           subtitle={EpageDescriptions.SALESMAN_DASHBOARD}
         />
-        <MeetingsByMonthChart />
-        <AvgDurationLineChart />
+        <Stack direction="row" spacing={2} sx={{ width: '100%' }}>
+          <Box sx={{ flex: 1, minWidth: 0, height: '100%' }}>
+            <MeetingsByMonthChart />
+          </Box>
+          <Box sx={{ flex: 1, minWidth: 0, height: '100%' }}>
+            <AvgDurationLineChart />
+          </Box>
+        </Stack>
       </Stack>
     </PageContainter>
   );
